+++ R/theme-defaults.r
@@ -188,6 +188,11 @@ theme_grey <- function(base_size = 11, base_family = "",
                            hjust = 1, vjust = 1,
                            margin = margin(t = half_line)
                          ),
+    plot.tag =           element_text(
+                           size = rel(1.3),
+                           hjust = 0, vjust = 1,
+                           margin = margin(r = half_line, b = half_line * 3)
+                         ),
     plot.margin =        margin(half_line, half_line, half_line, half_line),
 
     complete = TRUE
@@ -449,6 +454,11 @@ theme_void <- function(base_size = 11, base_family = "",
                            hjust = 1, vjust = 1,
                            margin = margin(t = half_line)
                          ),
+    plot.tag =           element_text(
+                           size = rel(1.3),
+                           hjust = 0, vjust = 1,
+                           margin = margin(r = half_line, b = half_line * 3)
+                         ),
 
     complete = TRUE
   )
@@ -565,6 +575,11 @@ theme_test <- function(base_size = 11, base_family = "",
                            size = rel(0.8),
                            hjust = 1, vjust = 1,
                            margin = margin(t = half_line)
+                         ),
+    plot.tag =           element_text(
+                           size = rel(1.3),
+                           hjust = 0, vjust = 1,
+                           margin = margin(r = half_line, b = half_line * 3)
                          ),
     plot.margin =        margin(half_line, half_line, half_line, half_line),